Split ondemand_object_tests into smaller files
